requirements.txt
python-dotenv == 1.0.1
langchain-huggingface == 0.1.2
tqdm == 4.67.0
transformers == 4.46.3
pandas == 2.2.3
accelerate == 1.1.1
datasets == 3.1.0
scikit-learn